Tagging the docker build and checking MakeDir in the tests
MakeDir was failing due to an incorrect 250 response from the ftpserver.  I reported this as a bug
and it's been fixed upstream.  The vendored package has been updated and the tests pass.  Also
tagging the docker build as bucketftp:latest but this will probably change to include the repo name.
We can do that once we know what repo we're pushing to.